+++ components/Sidebar/TopBar.tsx
@@ -40,11 +40,8 @@ export default function TopBar({ dark, on_click, icon_size, ...props }: TopBarPr
 	}, [scrollState])
 	const { isScrollingDown } = scrollState
 	const topBarClassName =
-		'flex fixed top-0 left-0 z-50 gap-6 \
-	items-center px-4 py-3 m-0 w-screen \
-	text-white shadow-lg sm:hidden \
-	bg-white-800 dark:text-dark \
-	dark:bg-discord-800 transition-all \
+		'flex fixed top-0 left-0 z-50 gap-6 items-center px-4 py-3 m-0 w-screen \
+	dark:text-white sm:hidden bg-white-800 dark:bg-discord-800 transition-all \
 	duration-300'
 
 	return (